requirements.txt
streamlit
transformers
torch
accelerate
sentencepiece
openai-whisper
TTS
soundfile
python-dotenv
openai
elevenlabs
